refactor: 패키지 이름 수정
- 파일 경로와 패키지명을 모델로 변경
