+++ gateway/src/main/java/ott/j4jg_gateway/model/dto/GoogleResponse.java
@@ -1,4 +1,4 @@
-package ott.j4jg_gateway.domain.dto;
+package ott.j4jg_gateway.model.dto;
 
 import java.util.Map;
