+++ gateway/src/main/java/ott/j4jg_gateway/model/dto/OAuth2Response.java
@@ -1,4 +1,4 @@
-package ott.j4jg_gateway.domain.dto;
+package ott.j4jg_gateway.model.dto;
 
 
 public interface OAuth2Response {
